Attempt to break the sound barrier
Applied optimizations on some computation-heavy functions: trying to
gain some precious milliseconds where possible using `numba`. I'm not
sure if this is really needed or even desirable but it is fairly easy to
remove if backtracking becomes a necessity.

Still trying to get Ableton Link Synchronization working properly. At
this point, I think that my implementation method is missing something
crucial. It is fairly hard to fall precisely in sync with Ableton.