Introducing... silence (for `S()`, `M()`, `O()`)
This is a pretty big update. It will surely break some things that will
need to be fixed pretty soon but it is important nonetheless to move
forward on this matter. This commit is adding the notion of silence into
patterns. This means that you can now write rhythms without having to
rely directly on `sleep()` or on patterning the recursive function.

There is a difference to be made between silence and parametric silence:
- if no sound, address or note is provided (depending on the situation
  and sender), the pattern will not emit any message. This is true
  silence.
- if a parameter is set to be silent, it will keep its current value for
  the duration of a silence. Checks are made recursively to ensure that
  it is possible to find a suitable value.

This commit is will ultimately make Sardine much more expressive and
enjoyable. Users will need to adjust to it..