Change samples in Sardinopedia Sample section
The existing sample files used in the Sample Position and Sample Stretching are not native to SuperDirt ('long').
I changed S sender to use 'fire' from the SD samples. It is a longer file that works well with different begin/end values and yields noticeable distortion with different timescale values. 

I also added a sentence in the SC Object that SC is only available when booting SC from Sardine.