Creating sardine.log if it doesn't exist!
A new file appeared without me noticing it! This file was implicitely
stating that the sardine.log exists whatever happens. This is not true,
especially for those who just installed Sardine. This commit should add
a verification process that will actually touch the file if it is not
there already.