Implementation of the logging lib
creation of an logger instance "log"  to print log in the rich console and
in an file located in the user folder.
The logging function is set-up to create file with a 1Mb size max.
beyond this limit, the file is back-up in the same place and a new file is created.
after 5 backed-up files. the oldest is overwrited.
For more information, cf. the documentation of logging.handlers.RotatingFileHandler

This logger is not yet propageted into the projet, and it log every message in console and in file
whitout discrimination.