logging : main.py
the majority of the message are intended to be show to the user.
As currently only the logging message beyond the error criticity will be display on the console and in the log file
so the message are print for the customer through logger.print (rich library) and also with logging.debug to save the informations.

Could be interesting to find a solution to avoid the duplication